Module Analyzer__.Alpha

type local_link = int * string

(link_id, link_name_for_debugging)

type a_arg =

| AFreeLink of string

| ALocalLink of local_link

| AAtom of string * a_arg list

type a_atom =

| ALocalInd of local_link * a_arg

| AFreeInd of string * a_arg

type indegs = (int * int) list * (string * int) list

(local_id, local_indeg) list, (free_name, free_indeg) list

type link_info = indegs * string list

(indegs, free_incidence list)

val alpha_arg : (string * int) list -> Parse.arg -> a_arg

A helper function for prep_proc Convert local link names to ids

val alpha_proc : (string * int) list -> int -> Parse.proc -> int * (a_atom, Parse.proc * Parse.proc) Stdlib.Either.t list

A helper function for alpha Convert local link names to ids and classify atoms and rules

val alpha : Parse.proc -> int * (a_atom list * (Parse.proc * Parse.proc) list)

Convert local link names to fresh ids and partition atoms and rules